+++ test/lib/Elastica/Test/StatusTest.php
@@ -7,10 +7,6 @@ use Elastica\Test\Base as BaseTest;
 
 class StatusTest extends BaseTest
 {
-    protected function setUp()
-    {
-        $this->es20();
-    }
 
     /**
      * @group functional
@@ -20,23 +16,6 @@ class StatusTest extends BaseTest
         $index = $this->_createIndex();
         $status = new Status($index->getClient());
         $this->assertInstanceOf('Elastica\Response', $status->getResponse());
-    }
-
-    /**
-     * @group functional
-     */
-    public function testGetIndexStatuses()
-    {
-        $index = $this->_createIndex();
-
-        $status = new Status($index->getClient());
-        $statuses = $status->getIndexStatuses();
-
-        $this->assertInternalType('array', $statuses);
-
-        foreach ($statuses as $indexStatus) {
-            $this->assertInstanceOf('Elastica\Index\Status', $indexStatus);
-        }
     }
 
     /**
@@ -115,24 +94,5 @@ class StatusTest extends BaseTest
             function ($index) {
                 return $index->getName();
             }, $indicesWithAlias));
-    }
-
-    /**
-     * @group functional
-     */
-    public function testServerStatus()
-    {
-        $client = $this->_getClient();
-        $status = $client->getStatus();
-        $serverStatus = $status->getServerStatus();
-
-        $this->assertTrue(!empty($serverStatus));
-        $this->assertTrue('array' == gettype($serverStatus));
-        $this->assertArrayHasKey('status', $serverStatus);
-        $this->assertTrue($serverStatus['status'] == 200);
-        $this->assertArrayHasKey('version', $serverStatus);
-
-        $versionInfo = $serverStatus['version'];
-        $this->assertArrayHasKey('number', $versionInfo);
     }
 }